Fixed two things I missed earlier
Forgot to remove :py27 from [testenv]

After that, I noticed the reason it was blowing up in python 3,
which was the end of the error output, was because wsgiref from
the requirements.txt actually failed to install on python3.

After closer inspection, it appeared wsgiref was not used anyway,
so I removed it from the requirements.txt

Things are working now, the tests are just not working on
python 3, but since you are switching to nose in another
branch anyway, I will leave that upto you.